requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
deepface==0.0.89
Pillow==10.1.0
tensorflow==2.15.0
tf-keras
mysql-connector-python==8.2.0
python-dotenv==1.0.0
requests==2.31.0
fastapi
uvicorn
python-multipart
slowapi==0.1.9
cachetools==5.3.2
loguru==0.7.2
psutil==5.9.6

